moved file playing and recording to a new file
